Initial scale gesture support on Android

+++ widget.hpp
@@ -22,7 +22,7 @@ namespace wheel
 		virtual void focus(widget *) {}
 		virtual point pointer() const { if(parent) return parent->pointer() - p; return point(); }
 		virtual void pointermove() { if(widget *c = at(pointer())) c->pointermove(); }
-		virtual void scroll(int d) { if(widget *c = focus()) c->scroll(d); }
+		virtual void scroll(float d) { if(widget *c = focus()) c->scroll(d); }
 		virtual void keycode(uint32_t k) { if(widget *c = focus()) c->keycode(k); }
 		virtual void press(uint8_t k) { if(widget *c = focus()) c->press(k); }
 		virtual void release(uint8_t k) { if(widget *c = focus()) c->release(k); }
